packages/justhtml/justhtml-1.4.0.tar.gz/justhtml-1.4.0/tests/justhtml-tests/branch_coverage.dat

#data
<math><mi>\x00</mi></math>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,11): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>


#data
<math><mi>\x00\x0c</mi></math>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,11): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         " "


#data
<math><mi>   </mi></math>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "   "


#data
</template>
#errors
(1,1): expected-doctype-but-got-end-tag
(1,11): unexpected-end-tag-before-html

#document
| <html>
|   <head>
|   <body>


#data
<body><input type="hidden">
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <input>
|       type="hidden"


#data
<html><body><html attr="test">
#errors
(1,1): expected-doctype-but-got-start-tag
(1,30): unexpected-start-tag

#document
| <html>
|   attr="test"
|   <head>
|   <body>


#data
<select><optgroup><option>1</option></optgroup></select>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|         <option>
|           "1"


#data
<select><option>1</option></optgroup></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "1"


#data
<!DOCTYPE html><html><head></head><frameset></frameset>
#errors

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>



#data
<!DOCTYPE html><html><head></head><frameset></frameset>abc
#errors
(1,58): unexpected-token-after-frameset

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>


#data
<table><input id="x" class="y" type="hidden"></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,45): unexpected-hidden-input-in-table

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         class="y"
|         id="x"
|         type="hidden"


#data
<math><annotation-xml encoding="text/plain"><div></div></annotation-xml></math>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,49): unexpected-html-element-in-foreign-content
(1,72): unexpected-end-tag
(1,79): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/plain"
|     <div>


#data
<math><annotation-xml><div></div></annotation-xml></math>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,27): unexpected-html-element-in-foreign-content
(1,50): unexpected-end-tag
(1,57): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|     <div>


#data
<svg><font color="red">text</font></svg>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,23): unexpected-html-element-in-foreign-content
(1,40): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <font>
|       color="red"
|       "text"


#data
<select><selectedcontent></selectedcontent><option>A</option><option selected>B</option></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,25): unexpected-start-tag-in-select
(1,43): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "B"
|       <option>
|         "A"
|       <option>
|         selected=""
|         "B"


#data
<select><selectedcontent></selectedcontent><option>A</option><option>B</option></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,25): unexpected-start-tag-in-select
(1,43): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "A"
|       <option>
|         "A"
|       <option>
|         "B"


#data
<b><select></b></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,15): unexpected-end-tag-in-select
(1,24): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <body>
|     <b>
|       <select>


#data
<a href="x"><select></a></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,24): unexpected-end-tag-in-select
(1,33): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="x"
|       <select>


#data
<select><plaintext></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,19): unexpected-start-tag-in-select
(1,28): expected-closing-tag-but-got-eof

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <plaintext>
|         "</select>"


#data
<svg><g><rect></rect><circle></circle><path></path></g></svg><div></div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         <svg rect>
|         <svg circle>
|         <svg path>
|     <div>


#data
<math><annotation-xml encoding="text/html"><div><table><tr></tr></table></div></annotation-xml></math>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <div>
|           <table>
|             <tbody>
|               <tr>


#data
<select></a></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,12): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<select></b></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,12): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<math><annotation-xml foo="bar"><div>test</div></annotation-xml></math>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,37): unexpected-html-element-in-foreign-content
(1,64): unexpected-end-tag
(1,71): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         foo="bar"
|     <div>
|       "test"


#data
<svg><g><rect><circle><path><div>text</div></path></circle></rect></g></svg>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,33): unexpected-html-element-in-foreign-content
(1,50): unexpected-end-tag
(1,59): unexpected-end-tag
(1,66): unexpected-end-tag
(1,70): unexpected-end-tag
(1,76): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         <svg rect>
|           <svg circle>
|             <svg path>
|     <div>
|       "text"


#data
<select><aside>text</aside></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,15): unexpected-start-tag-in-select
(1,27): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <select>
|       "text"


#data
<select><selectedcontent></selectedcontent><option disabled>A</option></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,25): unexpected-start-tag-in-select
(1,43): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "A"
|       <option>
|         disabled=""
|         "A"


#data
<select><selectedcontent></selectedcontent><option class="x">A</option><option selected>B</option></select>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,25): unexpected-start-tag-in-select
(1,43): unexpected-end-tag-in-select

#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "B"
|       <option>
|         class="x"
|         "A"
|       <option>
|         selected=""
|         "B"
